Extract Ownable mixin
