+++ app/controllers/api/templates_controller.rb
@@ -1,5 +1,6 @@
 module Api
   class TemplatesController < BaseController
+    include Shared::Ownable
     include Subcollections::Policies
     include Subcollections::PolicyProfiles
     include Subcollections::Tags
